+++ BLL/Models/Pin.cs
@@ -11,8 +11,8 @@ namespace BLL
         public int pin_id { get; set; }
         public string pin_title { get; set; }
         public string pin_content { get; set; }
-        public int pin_x { get; set; }
-        public int pin_y { get; set; }
+        public double pin_x { get; set; }
+        public double pin_y { get; set; }
         public int parent_map_id { get; set; }
     }
 }
